[#70] splitting the api test file in multiple files
